Makes the `from_url` method and the dependency on `reqwest` optional.
Introduces a new cargo feature "read-url". Reqwest is only a dependency
if "read-url" is enabled. This feature is enabled by default, so this is
not a breaking change: by default, nothing changes for downstream users.

Local building can be done by one of these options:
$ # As before:
$ cargo build
$ # Opting out of the default features, i.e., without method `from_url`:
$ cargo build --no-default-features
$ # Explicit, but equivalend to `cargo build`:
$ cargo build --no-default-features --features "read-url"

I also removed the dependency on the `regex` crate, since that doesn't
seem to be in use anyway.