Make (continuous) pickup drop off non-optional
This brings PickupDropOffType and ContinuousPickupDropOff in line with
other optional enums with default values like LocationType.
The GTFS specification makes pickup/dropoff optional, but gives a
default value on empty, so the Option type is misleading.
Additionally, this change allows parsing non-conforming values
(e.g. -999) and returning a default value instead.

This is a backwards incompatible change.