Add a new `RawGTFS::SourceFormat` field and strip file paths.
The new SourceFormat store the fact that the GTFS has been read from a
directory or a zip.

Also strip the base path in the `files` field of the `RawGTFS`. This
way, if the data were were from a directory with:
`RawGTFS::new("some/path/to/read")`, the `files` will contains
`stops.txt` and not `some/path/to/read/stops.txt`. It's more consistent
to the way it will be if the data is read from a zip (and makes it
possible to know if the data were in a subfolder).