Implement Fares v1
Implement Fares v1 as defined in https://gtfs.org/schedule/reference/#fare_attributestxt

There was already a FareAttribute, but no matching FareRule, so this
adds FareRule.

Fare v2 can also be implemented later if needed.

I did not try to be clever in the final `Gtfs` modelization as it is not
really needed for now, but maybe later we'll need to change the model a
bit.